factored bugsnag/logger config to its own file for import. wrote some basic fufactored bugsnag/logger into a separate file for  import. 4 functions to trigger different errors in the python terminal.
